backend/requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
databases==0.6.2
fastapi==0.87.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
